Teste de subtração com bugs

+++ Api/src/main/java/app/Subtracao.java
@@ -5,7 +5,7 @@ import io.jooby.annotations.*;
 @Path("/subtracao/{numUm}/{numDois}")
 public class Subtracao {
     @GET
-    public long calculaSubtracao(@PathParam int numUm, @PathParam int numDois) {
+    public long calculaSubtracao(@PathParam double numUm, @PathParam double numDois) {
         return numUm-numDois;
     }
 }
